Name or rename a list of gene trees or matrices

Name or rename a list of gene trees or gene matrices genes. For internal use mostly.

Usage

rename.genes(X, gene.names = NULL)

Arguments

X

A list of trees or matrices

gene.names

List of names to assign to the elements of X. Must be of the same length as length(X). If NULL (the default) the object are numbered 1,2,...,length(X).

Value

X with name assigned to each element.

Examples

data(carnivora)

# names before renaming
names(carnivora)
#>   [1] "ENSG00000004478_FKBP4"   "ENSG00000005381_MPO"    
#>   [3] "ENSG00000007944_MYLIP"   "ENSG00000012963_UBR7"   
#>   [5] "ENSG00000022355_GABRA1"  "ENSG00000023909_GCLM"   
#>   [7] "ENSG00000036473_OTC"     "ENSG00000048140_TSPAN17"
#>   [9] "ENSG00000049449_RCN1"    "ENSG00000051620_HEBP2"  
#>  [11] "ENSG00000054796_SPO11"   "ENSG00000069943_PIGB"   
#>  [13] "ENSG00000070010_UFD1"    "ENSG00000072954_TMEM38A"
#>  [15] "ENSG00000073111_MCM2"    "ENSG00000075336_TIMM21" 
#>  [17] "ENSG00000075975_MKRN2"   "ENSG00000079689_SCGN"   
#>  [19] "ENSG00000090432_MUL1"    "ENSG00000092010_PSME1"  
#>  [21] "ENSG00000092067_CEBPE"   "ENSG00000095380_NANS"   
#>  [23] "ENSG00000100078_PLA2G3"  "ENSG00000100591_AHSA1"  
#>  [25] "ENSG00000100604_CHGA"    "ENSG00000100625_SIX4"   
#>  [27] "ENSG00000102003_SYP"     "ENSG00000102245_CD40LG" 
#>  [29] "ENSG00000102543_CDADC1"  "ENSG00000104147_OIP5"   
#>  [31] "ENSG00000104221_BRF2"    "ENSG00000104679_R3HCC1" 
#>  [33] "ENSG00000105141_CASP14"  "ENSG00000105197_TIMM50" 
#>  [35] "ENSG00000105254_TBCB"    "ENSG00000105258_POLR2I" 
#>  [37] "ENSG00000106128_GHRHR"   "ENSG00000106511_MEOX2"  
#>  [39] "ENSG00000107815_TWNK"    "ENSG00000109072_VTN"    
#>  [41] "ENSG00000109471_IL2"     "ENSG00000110057_UNC93B1"
#>  [43] "ENSG00000110811_P3H3"    "ENSG00000110944_IL23A"  
#>  [45] "ENSG00000112096_SOD2"    "ENSG00000113302_IL12B"  
#>  [47] "ENSG00000113555_PCDH12"  "ENSG00000113734_BNIP1"  
#>  [49] "ENSG00000114388_NPRL2"   "ENSG00000114491_UMPS"   
#>  [51] "ENSG00000114638_UPK1B"   "ENSG00000114686_MRPL3"  
#>  [53] "ENSG00000114779_ABHD14B" "ENSG00000115665_SLC5A7" 
#>  [55] "ENSG00000116157_GPX7"    "ENSG00000116761_CTH"    
#>  [57] "ENSG00000117600_PLPPR4"  "ENSG00000118271_TTR"    
#>  [59] "ENSG00000119396_RAB14"   "ENSG00000120053_GOT1"   
#>  [61] "ENSG00000120659_TNFSF11" "ENSG00000121073_SLC35B1"
#>  [63] "ENSG00000121900_TMEM54"  "ENSG00000123307_NEUROD4"
#>  [65] "ENSG00000123999_INHA"    "ENSG00000124357_NAGK"   
#>  [67] "ENSG00000125538_IL1B"    "ENSG00000126353_CCR7"   
#>  [69] "ENSG00000127337_YEATS4"  "ENSG00000127993_RBM48"  
#>  [71] "ENSG00000128463_EMC4"    "ENSG00000128886_ELL3"   
#>  [73] "ENSG00000128965_CHAC1"   "ENSG00000130766_SESN2"  
#>  [75] "ENSG00000131013_PPIL4"   "ENSG00000132254_ARFIP2" 
#>  [77] "ENSG00000132275_RRP8"    "ENSG00000132463_GRSF1"  
#>  [79] "ENSG00000132693_CRP"     "ENSG00000133065_SLC41A1"
#>  [81] "ENSG00000133135_RNF128"  "ENSG00000133318_RTN3"   
#>  [83] "ENSG00000134107_BHLHE40" "ENSG00000134240_HMGCS2" 
#>  [85] "ENSG00000134588_USP26"   "ENSG00000134901_KDELC1" 
#>  [87] "ENSG00000135452_TSPAN31" "ENSG00000135469_COQ10A" 
#>  [89] "ENSG00000136156_ITM2B"   "ENSG00000136634_IL10"   
#>  [91] "ENSG00000136807_CDK9"    "ENSG00000136932_TRMO"   
#>  [93] "ENSG00000136999_NOV"     "ENSG00000138231_DBR1"   
#>  [95] "ENSG00000138675_FGF5"    "ENSG00000139572_GPR84"  
#>  [97] "ENSG00000140274_DUOXA2"  "ENSG00000140450_ARRDC4" 
#>  [99] "ENSG00000140459_CYP11A1" "ENSG00000141200_KIF2B"  
#> [101] "ENSG00000141349_G6PC3"   "ENSG00000141753_IGFBP4" 
#> [103] "ENSG00000142039_CCDC97"  "ENSG00000142733_MAP3K6" 
#> [105] "ENSG00000142871_CYR61"   "ENSG00000143125_PROK1"  
#> [107] "ENSG00000143153_ATP1B1"  "ENSG00000143196_DPT"    
#> [109] "ENSG00000143942_CHAC2"   "ENSG00000144136_SLC20A1"
#> [111] "ENSG00000144355_DLX1"    "ENSG00000145050_MANF"   
#> [113] "ENSG00000145384_FABP2"   "ENSG00000145391_SETD7"  
#> [115] "ENSG00000145649_GZMA"    "ENSG00000147457_CHMP7"  
#> [117] "ENSG00000147679_UTP23"   "ENSG00000148229_POLE3"  
#> [119] "ENSG00000148331_ASB6"    "ENSG00000148334_PTGES2" 
#> [121] "ENSG00000148677_ANKRD1"  "ENSG00000148795_CYP17A1"
#> [123] "ENSG00000149243_KLHL35"  "ENSG00000149573_MPZL2"  
#> [125] "ENSG00000151611_MMAA"   

carnivora.renamed<-rename.genes(carnivora, gene.names=as.character(1:length(carnivora)))

# names after renaming
names(carnivora.renamed)
#>   [1] "1"   "2"   "3"   "4"   "5"   "6"   "7"   "8"   "9"   "10"  "11"  "12" 
#>  [13] "13"  "14"  "15"  "16"  "17"  "18"  "19"  "20"  "21"  "22"  "23"  "24" 
#>  [25] "25"  "26"  "27"  "28"  "29"  "30"  "31"  "32"  "33"  "34"  "35"  "36" 
#>  [37] "37"  "38"  "39"  "40"  "41"  "42"  "43"  "44"  "45"  "46"  "47"  "48" 
#>  [49] "49"  "50"  "51"  "52"  "53"  "54"  "55"  "56"  "57"  "58"  "59"  "60" 
#>  [61] "61"  "62"  "63"  "64"  "65"  "66"  "67"  "68"  "69"  "70"  "71"  "72" 
#>  [73] "73"  "74"  "75"  "76"  "77"  "78"  "79"  "80"  "81"  "82"  "83"  "84" 
#>  [85] "85"  "86"  "87"  "88"  "89"  "90"  "91"  "92"  "93"  "94"  "95"  "96" 
#>  [97] "97"  "98"  "99"  "100" "101" "102" "103" "104" "105" "106" "107" "108"
#> [109] "109" "110" "111" "112" "113" "114" "115" "116" "117" "118" "119" "120"
#> [121] "121" "122" "123" "124" "125"